ensures player is always the first thing
previously the lua code would find the player in the list of things. this causes issues because it wouldn't update the stored index if the player's index changed (e.g. by picking up an item earlier in the things list). that can be fixed by making the player always first